refactor(design-tokens): group slot targets under a Slot sub-namespace
Move the slot-target family into Projection\Slot: the Slot_Target contract
becomes Slot\Contracts\Target, and the Kb_ prefix drops from the
implementations (Spacing_Target, Gap_Target). The standalone targets
(Wp_Preset_Target, Kadence_Palette_Slot) stay flat at Projection\ as before;
only the contract-backed family gets its own namespace, mirroring Adapter\.
This also retires the single-file Projection\Contracts directory.
